+++ src/OctoMap/Abstractions/IOctoProjectionBuilder.cs
@@ -16,11 +16,29 @@ namespace OctoMap
         Expression<Func<TSource, TDestination>> Build<TSource, TDestination>();
 
         /// <summary>
+        /// Builds a projection expression for the specified source and destination types using external parameters.
+        /// </summary>
+        /// <typeparam name="TSource">The source type.</typeparam>
+        /// <typeparam name="TDestination">The destination type.</typeparam>
+        /// <param name="parameters">The projection parameters.</param>
+        /// <returns>The projection expression.</returns>
+        Expression<Func<TSource, TDestination>> Build<TSource, TDestination>(object parameters);
+
+        /// <summary>
         /// Builds a projection expression for the specified source and destination types.
         /// </summary>
         /// <param name="sourceType">The source type.</param>
         /// <param name="destinationType">The destination type.</param>
         /// <returns>The projection expression.</returns>
         LambdaExpression Build(Type sourceType, Type destinationType);
+
+        /// <summary>
+        /// Builds a projection expression for the specified source and destination types using external parameters.
+        /// </summary>
+        /// <param name="sourceType">The source type.</param>
+        /// <param name="destinationType">The destination type.</param>
+        /// <param name="parameters">The projection parameters.</param>
+        /// <returns>The projection expression.</returns>
+        LambdaExpression Build(Type sourceType, Type destinationType, object parameters);
     }
 }